Move verifier cmd.
