+++ cmd/verifier/committee/air.toml
@@ -1,8 +1,8 @@
-root = ".."
+root = "../../.."
 
 [build]
-cmd = "go build -o ./tmp/token /app/verifier/cmd/token"
-bin = "tmp/token"
+cmd = "go build -o ./tmp/committee /app/cmd/verifier/committee"
+bin = "tmp/committee"
 include_dir = ["app"]
 include_ext = ["go", "mod", "sum"]
 exclude_dir = ["tmp", "vendor", "testdata", ".git"]
