+++ cmd/verifier/common.go
