+++ cmd/verifier/token/main.go
@@ -19,7 +19,7 @@ import (
 	"github.com/smartcontractkit/chainlink-ccv/pkg/chainaccess"
 	"github.com/smartcontractkit/chainlink-ccv/protocol"
 	"github.com/smartcontractkit/chainlink-ccv/protocol/common/logging"
-	cmd "github.com/smartcontractkit/chainlink-ccv/verifier/cmd"
+	cmd "github.com/smartcontractkit/chainlink-ccv/cmd/verifier"
 	verifier "github.com/smartcontractkit/chainlink-ccv/verifier/pkg"
 	"github.com/smartcontractkit/chainlink-ccv/verifier/pkg/chainstatus"
 	"github.com/smartcontractkit/chainlink-ccv/verifier/pkg/monitoring"
