+++ verifier/Dockerfile
@@ -11,7 +11,7 @@ RUN --mount=type=cache,target=/go/pkg/mod,id=ccv-go-mod \
 COPY . .
 
 ARG VERIFIER_TYPE=committee
-WORKDIR /verifier/cmd/${VERIFIER_TYPE}
+WORKDIR /cmd/verifier/${VERIFIER_TYPE}
 
 RUN --mount=type=cache,target=/root/.cache/go-build,id=ccv-go-build \
     --mount=type=cache,target=/go/pkg/mod,id=ccv-go-mod \
